+++ packages/nde-erfgoed-core/lib/collections/collection-memory-store.ts
@@ -32,4 +32,10 @@ export class CollectionMemoryStore extends MemoryStore<Collection> implements Co
 
   }
 
+  async getCollection(uri: string): Promise<Collection> {
+
+    return this.resources.find((res) => res.uri === uri);
+
+  }
+
 }