+++ packages/nde-erfgoed-core/lib/collections/collection-object-store.ts
@@ -6,6 +6,7 @@ import { CollectionObject } from './collection-object';
  * A store for collection objects.
  */
 export interface CollectionObjectStore extends Store<CollectionObject> {
+
   /**
    * Retrieves all objects for a specific collection.
    *
@@ -19,4 +20,12 @@ export interface CollectionObjectStore extends Store<CollectionObject> {
    * @param searchTerm the therm that should be present in a property
    */
   search(searchTerm: string): Promise<CollectionObject[]>;
+
+  /**
+   * Retrieves a single CollectionObject
+   *
+   * @param uri The URI of the CollectionObject
+   */
+  getObject(uri: string): Promise<CollectionObject>;
+
 }