+++ packages/nde-erfgoed-manage/lib/features/search/search-root.component.ts
@@ -192,17 +192,17 @@ export class SearchRootComponent extends RxLitElement {
         }
         .grid {
           display: grid;
-          grid-template-columns: repeat(3, 1fr);
+          grid-template-columns: repeat(3, minmax(0, 1fr));
           grid-gap: var(--gap-large);
         }
         @media only screen and (max-width: 1300px) {
           .grid {
-            grid-template-columns: repeat(2, 1fr);
+            grid-template-columns: repeat(2, minmax(0, 1fr));
           }
         }
         @media only screen and (max-width: 1000px) {
           .grid {
-            grid-template-columns: repeat(1, 1fr);
+            grid-template-columns: repeat(1, minmax(0, 1fr));
           }
         }
         nde-object-card, nde-collection-card {