+++ packages/nde-erfgoed-theme/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.0](https://github.com/digita-ai/nde-erfgoedinstellingen/compare/v0.2.1...v0.3.0) (2021-05-18)
+
+
+### **Features**
+
+* support updating of collections ([#189](https://github.com/digita-ai/nde-erfgoedinstellingen/issues/189)) ([4e8edc2](https://github.com/digita-ai/nde-erfgoedinstellingen/commit/4e8edc2243070a48042dd3a8af4a29ab86058ed4))
+
+
+
+### [0.2.1](https://github.com/digita-ai/nde-erfgoedinstellingen/compare/v0.2.0...v0.2.1) (2021-05-11)
+
+**Note:** Version bump only for package @digita-ai/nde-erfgoed-theme
+
+
+
+
+
 ## 0.2.0 (2021-05-10)